Rename RpcState to SessionState
